+++ decitala/path_finding/dijkstra.py
@@ -7,45 +7,21 @@
 #
 # Location: NYC, 2021
 ####################################################################################################
-"""
-function Dijkstra(Graph, source):
- 2
- 3      create vertex set Q
- 4
- 5      for each vertex v in Graph:            
- 6          dist[v] ← INFINITY                 
- 7          prev[v] ← UNDEFINED                
- 8          add v to Q                     
- 9      dist[source] ← 0                       
-10     
-11      while Q is not empty:
-12          u ← vertex in Q with min dist[u]   
-13                                             
-14          remove u from Q
-15         
-16          for each neighbor v of u:           // only v that are still in Q
-17              alt ← dist[u] + length(u, v)
-18              if alt < dist[v]:              
-19                  dist[v] ← alt
-20                  prev[v] ← u
-21
-22      return dist[], prev[]
-
-If we are only interested in a shortest path between vertices source and target, we can terminate the search after line 15 if u = target. Now we can read the shortest path from source to target by reverse iteration:
-
-1  S ← empty sequence
-2  u ← target
-3  if prev[u] is defined or u = source:          // Do something only if the vertex is reachable
-4      while u is defined:                       // Construct the shortest path with a stack S
-5          insert u at the beginning of S        // Push the vertex onto the stack
-6          u ← prev[u]                           // Traverse from target to source
-"""
 import numpy as np
 
-def dijkstra(data, source, weights, verbose=False):
+from .path_finding_utils import cost
+
+def dijkstra(
+		data,
+		source,
+		target,
+		weights,
+		verbose=False
+	):
 	vertices = []
 	dist = [np.inf] * len(data)
 	prev = [None] * len(data)
+	
 	source_index = 0
 	for i, fragment_data in enumerate(data):
 		if fragment_data == source:
@@ -55,7 +31,18 @@ def dijkstra(data, source, weights, verbose=False):
 	dist[source_index] = 0
 
 	while vertices:
-		# u ← vertex in Q with min dist[u]
+		curr_vertex = min(dist)
+		del vertices[dist.index(u)]
+
+		if u == target:
+			break
+
+		for v in vertices:
+			alt = dist[u] + cost(u, v, weights=weights)
+			# if alt < dist[v]:
+			# 	dist[v] = alt
+			# 	prev[v] = u
+		
 		break
 
 	return dist, prev